Adds Stanley's changes for 0F30 but not 0F10
There is some confusion at the moment. The change at 0F30 seems to be
a valid change and has been made. But the's a difference in the
manuals for the 0F10 changes. My manual is pretty clean on 0F10. But
my manual also doesn't have the 0F30 S19 record. Hence the confusion.