Updates inc to allow for an EXP option and remove redundant macros
I've modified the inc file to allow an EXP option (-D EXP). I also
moved the messages outside the 3 options (ROMORIG, EXP and the
default). The default still need to be properly setup so the code can
be dropped in a 2732 ROM that sits below $F000 (MIKBUG) and using the
Microware addressing modifications.

Changes to be committed:
	modified:   rt68mx.inc